+++ src/darker/help.py
@@ -21,11 +21,12 @@ DESCRIPTION = "".join(DESCRIPTION_PARTS)
 SRC = "Path(s) to the Python source file(s) to reformat"
 
 REVISION = (
-    "Git revision against which to compare the working tree. Tags, branch"
-    " names, commit hashes, and other expressions like HEAD~5 work here. Also"
-    " a range like master...HEAD or master... can be used to compare the best"
-    " common ancestor. With the magic value :PRE-COMMIT:, Darker expects the revision"
-    " range from the PRE_COMMIT_FROM_REF and PRE_COMMIT_TO_REF environment variables."
+    "Git revision against which to compare the working tree. Tags, branch names, commit"
+    " hashes, and other expressions like HEAD~5 work here. Also a range like"
+    " master...HEAD or master... can be used to compare the best common ancestor. With"
+    " the magic value :PRE-COMMIT:, Darker works in pre-commit compatible mode. Darker"
+    " expects the revision range from the PRE_COMMIT_FROM_REF and PRE_COMMIT_TO_REF"
+    " environment variables. If those are not found, Darker works against HEAD."
 )
 
 DIFF = (